Fill the third tab with sliders and inputs
It adds the play button and the forward and reverse buttons.

It adds Input Box and Sliders for the Number of Congressional Districts
and Number of Max-Min Districts.

It adds a dropdown menu to select Ethnic Groups.

It adds a range slider to change the Min-Max percentages of the ethnic
groups.